# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/email/config:
        post:
            tags:
                - Email
            operationId: Email_CreateEmailConfig
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.CreateEmailConfigRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.CreateEmailConfigReply'
    /v1/email/config/{uid}:
        get:
            tags:
                - Email
            operationId: Email_GetEmailConfig
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.EmailConfigItem'
        put:
            tags:
                - Email
            operationId: Email_UpdateEmailConfig
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.UpdateEmailConfigRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.UpdateEmailConfigReply'
        delete:
            tags:
                - Email
            operationId: Email_DeleteEmailConfig
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.DeleteEmailConfigReply'
    /v1/email/config/{uid}/status:
        put:
            tags:
                - Email
            operationId: Email_UpdateEmailConfigStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.UpdateEmailConfigStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.UpdateEmailConfigStatusReply'
    /v1/email/configs:
        get:
            tags:
                - Email
            operationId: Email_ListEmailConfig
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.ListEmailConfigReply'
    /v1/email/configs/select:
        get:
            tags:
                - Email
            operationId: Email_SelectEmailConfig
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SelectEmailConfigReply'
...
    /v1/sender/email/{uid}:
        post:
            tags:
                - Sender
            operationId: Sender_SendEmail
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendEmailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
...
components:
    schemas:
...
        rabbit.api.v1.CreateEmailConfigReply:
            type: object
            properties: {}
        rabbit.api.v1.CreateEmailConfigRequest:
            type: object
            properties:
                name:
                    type: string
                host:
                    type: string
                port:
                    type: integer
                    format: int32
                username:
                    type: string
                password:
                    type: string
...
        rabbit.api.v1.DeleteEmailConfigReply:
            type: object
            properties: {}
...
        rabbit.api.v1.EmailConfigItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                host:
                    type: string
                port:
                    type: integer
                    format: int32
                username:
                    type: string
                password:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                status:
                    type: integer
                    format: enum
        rabbit.api.v1.EmailConfigItemSelect:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
...
        rabbit.api.v1.ListEmailConfigReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.EmailConfigItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        rabbit.api.v1.SelectEmailConfigReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.EmailConfigItemSelect'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        rabbit.api.v1.SendEmailRequest:
            type: object
            properties:
                uid:
                    type: string
                subject:
                    type: string
                body:
                    type: string
                to:
                    type: array
                    items:
                        type: string
                cc:
                    type: array
                    items:
                        type: string
                contentType:
                    type: string
                headers:
                    type: object
                    additionalProperties:
                        type: string
...
        rabbit.api.v1.SendReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
...
        rabbit.api.v1.UpdateEmailConfigReply:
            type: object
            properties: {}
        rabbit.api.v1.UpdateEmailConfigRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                host:
                    type: string
                port:
                    type: integer
                    format: int32
                username:
                    type: string
                password:
                    type: string
        rabbit.api.v1.UpdateEmailConfigStatusReply:
            type: object
            properties: {}
        rabbit.api.v1.UpdateEmailConfigStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
...