# pkg/api/v1/email_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/v1/email.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationEmailCreateEmailConfig = "/rabbit.api.v1.Email/CreateEmailConfig"
const OperationEmailDeleteEmailConfig = "/rabbit.api.v1.Email/DeleteEmailConfig"
const OperationEmailGetEmailConfig = "/rabbit.api.v1.Email/GetEmailConfig"
const OperationEmailListEmailConfig = "/rabbit.api.v1.Email/ListEmailConfig"
const OperationEmailSelectEmailConfig = "/rabbit.api.v1.Email/SelectEmailConfig"
const OperationEmailUpdateEmailConfig = "/rabbit.api.v1.Email/UpdateEmailConfig"
const OperationEmailUpdateEmailConfigStatus = "/rabbit.api.v1.Email/UpdateEmailConfigStatus"

type EmailHTTPServer interface {
	CreateEmailConfig(context.Context, *CreateEmailConfigRequest) (*CreateEmailConfigReply, error)
	DeleteEmailConfig(context.Context, *DeleteEmailConfigRequest) (*DeleteEmailConfigReply, error)
	GetEmailConfig(context.Context, *GetEmailConfigRequest) (*EmailConfigItem, error)
	ListEmailConfig(context.Context, *ListEmailConfigRequest) (*ListEmailConfigReply, error)
	SelectEmailConfig(context.Context, *SelectEmailConfigRequest) (*SelectEmailConfigReply, error)
	UpdateEmailConfig(context.Context, *UpdateEmailConfigRequest) (*UpdateEmailConfigReply, error)
	UpdateEmailConfigStatus(context.Context, *UpdateEmailConfigStatusRequest) (*UpdateEmailConfigStatusReply, error)
}

func RegisterEmailHTTPServer(s *http.Server, srv EmailHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/email/config", _Email_CreateEmailConfig0_HTTP_Handler(srv))
	r.PUT("/v1/email/config/{uid}", _Email_UpdateEmailConfig0_HTTP_Handler(srv))
	r.PUT("/v1/email/config/{uid}/status", _Email_UpdateEmailConfigStatus0_HTTP_Handler(srv))
	r.DELETE("/v1/email/config/{uid}", _Email_DeleteEmailConfig0_HTTP_Handler(srv))
	r.GET("/v1/email/config/{uid}", _Email_GetEmailConfig0_HTTP_Handler(srv))
	r.GET("/v1/email/configs", _Email_ListEmailConfig0_HTTP_Handler(srv))
	r.GET("/v1/email/configs/select", _Email_SelectEmailConfig0_HTTP_Handler(srv))
}

func _Email_CreateEmailConfig0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateEmailConfigRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailCreateEmailConfig)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateEmailConfig(ctx, req.(*CreateEmailConfigRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateEmailConfigReply)
		return ctx.Result(200, reply)
	}
}

func _Email_UpdateEmailConfig0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateEmailConfigRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailUpdateEmailConfig)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateEmailConfig(ctx, req.(*UpdateEmailConfigRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateEmailConfigReply)
		return ctx.Result(200, reply)
	}
}

func _Email_UpdateEmailConfigStatus0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateEmailConfigStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailUpdateEmailConfigStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateEmailConfigStatus(ctx, req.(*UpdateEmailConfigStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateEmailConfigStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Email_DeleteEmailConfig0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteEmailConfigRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailDeleteEmailConfig)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteEmailConfig(ctx, req.(*DeleteEmailConfigRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteEmailConfigReply)
		return ctx.Result(200, reply)
	}
}

func _Email_GetEmailConfig0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetEmailConfigRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailGetEmailConfig)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetEmailConfig(ctx, req.(*GetEmailConfigRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*EmailConfigItem)
		return ctx.Result(200, reply)
	}
}

func _Email_ListEmailConfig0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListEmailConfigRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailListEmailConfig)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListEmailConfig(ctx, req.(*ListEmailConfigRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListEmailConfigReply)
		return ctx.Result(200, reply)
	}
}

func _Email_SelectEmailConfig0_HTTP_Handler(srv EmailHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectEmailConfigRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationEmailSelectEmailConfig)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectEmailConfig(ctx, req.(*SelectEmailConfigRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectEmailConfigReply)
		return ctx.Result(200, reply)
	}
}

type EmailHTTPClient interface {
	CreateEmailConfig(ctx context.Context, req *CreateEmailConfigRequest, opts ...http.CallOption) (rsp *CreateEmailConfigReply, err error)
	DeleteEmailConfig(ctx context.Context, req *DeleteEmailConfigRequest, opts ...http.CallOption) (rsp *DeleteEmailConfigReply, err error)
	GetEmailConfig(ctx context.Context, req *GetEmailConfigRequest, opts ...http.CallOption) (rsp *EmailConfigItem, err error)
	ListEmailConfig(ctx context.Context, req *ListEmailConfigRequest, opts ...http.CallOption) (rsp *ListEmailConfigReply, err error)
	SelectEmailConfig(ctx context.Context, req *SelectEmailConfigRequest, opts ...http.CallOption) (rsp *SelectEmailConfigReply, err error)
	UpdateEmailConfig(ctx context.Context, req *UpdateEmailConfigRequest, opts ...http.CallOption) (rsp *UpdateEmailConfigReply, err error)
	UpdateEmailConfigStatus(ctx context.Context, req *UpdateEmailConfigStatusRequest, opts ...http.CallOption) (rsp *UpdateEmailConfigStatusReply, err error)
}

type EmailHTTPClientImpl struct {
	cc *http.Client
}

func NewEmailHTTPClient(client *http.Client) EmailHTTPClient {
	return &EmailHTTPClientImpl{client}
}

func (c *EmailHTTPClientImpl) CreateEmailConfig(ctx context.Context, in *CreateEmailConfigRequest, opts ...http.CallOption) (*CreateEmailConfigReply, error) {
	var out CreateEmailConfigReply
	pattern := "/v1/email/config"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationEmailCreateEmailConfig))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *EmailHTTPClientImpl) DeleteEmailConfig(ctx context.Context, in *DeleteEmailConfigRequest, opts ...http.CallOption) (*DeleteEmailConfigReply, error) {
	var out DeleteEmailConfigReply
	pattern := "/v1/email/config/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationEmailDeleteEmailConfig))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *EmailHTTPClientImpl) GetEmailConfig(ctx context.Context, in *GetEmailConfigRequest, opts ...http.CallOption) (*EmailConfigItem, error) {
	var out EmailConfigItem
	pattern := "/v1/email/config/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationEmailGetEmailConfig))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *EmailHTTPClientImpl) ListEmailConfig(ctx context.Context, in *ListEmailConfigRequest, opts ...http.CallOption) (*ListEmailConfigReply, error) {
	var out ListEmailConfigReply
	pattern := "/v1/email/configs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationEmailListEmailConfig))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *EmailHTTPClientImpl) SelectEmailConfig(ctx context.Context, in *SelectEmailConfigRequest, opts ...http.CallOption) (*SelectEmailConfigReply, error) {
	var out SelectEmailConfigReply
	pattern := "/v1/email/configs/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationEmailSelectEmailConfig))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *EmailHTTPClientImpl) UpdateEmailConfig(ctx context.Context, in *UpdateEmailConfigRequest, opts ...http.CallOption) (*UpdateEmailConfigReply, error) {
	var out UpdateEmailConfigReply
	pattern := "/v1/email/config/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationEmailUpdateEmailConfig))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *EmailHTTPClientImpl) UpdateEmailConfigStatus(ctx context.Context, in *UpdateEmailConfigStatusRequest, opts ...http.CallOption) (*UpdateEmailConfigStatusReply, error) {
	var out UpdateEmailConfigStatusReply
	pattern := "/v1/email/config/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationEmailUpdateEmailConfigStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}